requirements.txt
Flask == 3.1.0
torch == 2.7.0
torchvision == 0.22.0
Pillow == 11.0.0
